+++ Umpah-iOS/Umpah-iOS/Screen/Main/Controller/MainVC.swift
@@ -13,53 +13,14 @@ import RxSwift
 import SnapKit
 import Then
 
-final class MainVC: BaseViewController {
+final class MainVC: MainTableVC {
 
     // MARK: - properties
     
-    public lazy var baseTableView = UITableView().then {
-        $0.delegate = self
-        $0.dataSource = self
-        $0.register(ChartTVC.self, forCellReuseIdentifier: ChartTVC.identifier)
-        $0.register(DetailTVC.self, forCellReuseIdentifier: DetailTVC.identifier)
-        $0.register(FilterTVC.self, forCellReuseIdentifier: FilterTVC.identifier)
-        $0.register(StrokeTVC.self, forCellReuseIdentifier: StrokeTVC.identifier)
-        $0.register(DateTVC.self, forCellReuseIdentifier: DateTVC.identifier)
-        $0.register(RoutineTVC.self)
-        $0.backgroundColor = .clear
-        $0.estimatedRowHeight = 100
-        $0.separatorStyle = .none
-        $0.showsVerticalScrollIndicator = false
-        
-        if #available(iOS 15.0, *) {
-            $0.sectionHeaderTopPadding = 0
-        }
-    }
-    public lazy var cardView = CardView(rootVC: self)
-    
-    private let topView = TopView()
-    private let headerView = HeaderView()
-    private let statusBar = StatusBar()
-    private lazy var dateText: String = dateformatter.string(from: Date())
-    private lazy var rangeTexts: [String] = [serverDateformatter.string(from: Date()), ""]
-    private var cardViewTopConstraint: NSLayoutConstraint?
-    private var cardPanStartingTopConstant : CGFloat = 20.0
-    private var cardPanMaxVelocity: CGFloat = 1500.0
-    private var currentState: CurrentState = .base
-    private var cacheState: CurrentState = .base
-    private var strokeState: Stroke = .none
-    private var dateformatter = DateFormatter().then {
-        $0.dateFormat = "YY/MM/dd"
-        $0.locale = Locale.init(identifier: "ko-KR")
-    }
-    private var serverDateformatter = DateFormatter().then {
-        $0.dateFormat = "YYYY-MM-dd"
-        $0.locale = Locale.init(identifier: "ko-KR")
-    }
 
-    var routineOverViewList: [RoutineOverviewData] = []
+    
     let swimmingViewModel = SwimmingDataViewModel()
-    let storage = RecordStorage.shared
+    
     
     // MARK: - life cycle
     
@@ -67,6 +28,8 @@ final class MainVC: BaseViewController {
         super.viewDidLoad()
         addClosureToChangeState()
         authorizeHealthKit()
+        
+        baseTableView.delegate = self
     }
     
     override func viewDidAppear(_ animated: Bool) {
@@ -111,10 +74,7 @@ final class MainVC: BaseViewController {
         initGestureView()
         initRoutineOverViewList()
     }
-}
-
-// MARK: - Custom Methods
-extension MainVC {
+    
     private func initGestureView() {
         let viewPan = UIPanGestureRecognizer(target: self, action: #selector(viewPanned(_:)))
         viewPan.delaysTouchesBegan = false
@@ -146,173 +106,48 @@ extension MainVC {
     }
 }
 
-// MARK: - Animation
+// MARK: - HealthKit
 extension MainVC {
-    // MARK: - @objc
-    
-    @objc
-    func viewPanned(_ panRecognizer: UIPanGestureRecognizer) {
-        let velocity = panRecognizer.velocity(in: self.view)
-        let translation = panRecognizer.translation(in: self.view)
-        
-        if cardView.expandedView.isModified {
-            panRecognizer.state = .failed
-        }
-        
-        switch panRecognizer.state {
-        case .began:
-            cardPanStartingTopConstant = cardViewTopConstraint?.constant ?? 0
-        case .changed:
-            if self.cardPanStartingTopConstant + translation.y > 20.0 {
-                self.cardViewTopConstraint?.constant = self.cardPanStartingTopConstant + translation.y
-            }
-        case .ended:
-            if velocity.y > cardPanMaxVelocity {
-                cardView.cardViewState = .normal
-                decideTopConstraint(of: .normal)
+    private func authorizeHealthKit() {
+        HealthKitSetupAssistant.authorizeHealthKitAtSwimming { (authorized, error) in
+            guard authorized else {
+                let baseMessage = "HealthKit Authorization Failed"
+                if let error = error {
+                    print("\(baseMessage). Reason: \(error.localizedDescription)")
+                } else {
+                    print(baseMessage)
+                }
                 return
             }
-            
-            let window = UIApplication.shared.windows.first { $0.isKeyWindow }
-            if let safeAreaHeight = window?.safeAreaLayoutGuide.layoutFrame.size.height,
-               let bottomPadding = window?.safeAreaInsets.bottom {
-                if self.cardViewTopConstraint?.constant ?? 0 < (safeAreaHeight + bottomPadding) * 0.6 {
-                    cardView.cardViewState = .expanded
-                    decideTopConstraint(of: .expanded)
-                } else  {
-                    cardView.cardViewState = .normal
-                    decideTopConstraint(of: .normal)
-                }
+            print("HealthKit Successfully Authorized.")
+            self.postRecordData()
+            self.swimmingViewModel.initSwimmingData()
+            DispatchQueue.main.asyncAfter(deadline: .now() + 3){
+                self.swimmingViewModel.refineSwimmingDataForServer()
+                print("refineSwimmingDataForServer.")
             }
-        default:
-            cardView.cardViewState = .fail
-        }
-    }
-    
-    public func decideTopConstraint(of state: CardViewState) {
-        switch state {
-        case .normal:
-            cardViewTopConstraint?.constant = UIScreen.main.hasNotch ? UIScreen.main.bounds.size.height * 0.8 : UIScreen.main.bounds.size.height * 0.87
-        case .expanded:
-            switch currentState {
-            case .week:
-                cardViewTopConstraint?.constant = UIScreen.main.bounds.size.height * 0.24
-            case .month:
-                cardViewTopConstraint?.constant = UIScreen.main.bounds.size.height * 0.38
-            default:
-                cardViewTopConstraint?.constant = 20.0
-            }
-        case .fail:
-            break
         }
     }
 }
 
-// MARK: - UITableViewDataSource
-extension MainVC: UITableViewDataSource {
-    func numberOfSections(in tableView: UITableView) -> Int {
-        return 2
-    }
-
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        switch section {
-        case 1:
-            if currentState != .routine {
-                return 5
-            } else {
-                return routineOverViewList.count
-            }
-        default:
-            return 0
-        }
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        switch currentState {
-        case .day,
-             .base:
-            switch indexPath.row {
-            case 0:
-                guard let cell = tableView.dequeueReusableCell(withIdentifier: FilterTVC.identifier) as? FilterTVC else { return UITableViewCell() }
-                cell.delegate = self
-                cell.state = currentState
-                return cell
-            case 1:
-                guard let cell = tableView.dequeueReusableCell(withIdentifier: DateTVC.identifier) as? DateTVC else { return UITableViewCell() }
-                cell.dateLabel.text = dateText
-                cell.dateLabel.addCharacterSpacing(kernValue: 2)
-                cell.dateLabel.font = .nexaBold(ofSize: 16)
-                return cell
-            case 2:
-                guard let cell = tableView.dequeueReusableCell(withIdentifier: DetailTVC.identifier) as? DetailTVC else { return UITableViewCell() }
-                cell.titleLabel.text = "OVERVIEW"
-                cell.titleLabel.addCharacterSpacing(kernValue: 2)
-                return cell
-            case 3:
-                guard let cell = tableView.dequeueReusableCell(withIdentifier: StrokeTVC.identifier) as? StrokeTVC else { return UITableViewCell() }
-                cell.titleLabel.text = "TOTAL"
-                cell.titleLabel.addCharacterSpacing(kernValue: 2)
-                return cell
-            default:
-                let cell = UITableViewCell(frame: .zero)
-                cell.backgroundColor = .upuhBackground
-                cell.selectionStyle = .none
-                return cell
-            }
-        case .week,
-             .month:
-            switch indexPath.row {
-            case 0:
-                guard let cell = tableView.dequeueReusableCell(withIdentifier: FilterTVC.identifier) as? FilterTVC else { return UITableViewCell() }
-                cell.delegate = self
-                cell.state = currentState
-                cell.stroke = strokeState
-                return cell
-            case 1:
-                guard let cell = tableView.dequeueReusableCell(withIdentifier: DateTVC.identifier) as? DateTVC else { return UITableViewCell() }
-                cell.dateLabel.text = dateText
-                cell.dateLabel.addCharacterSpacing(kernValue: 2)
-                if dateText == "이번주" || dateText == "지난주" {
-                    cell.dateLabel.font = .IBMPlexSansSemiBold(ofSize: 16)
-                } else {
-                    cell.dateLabel.font = .nexaBold(ofSize: 16)
-                }
-                return cell
-            case 2:
-                guard let cell = tableView.dequeueReusableCell(withIdentifier: ChartTVC.identifier) as? ChartTVC else { return UITableViewCell() }
-                cell.combineChartView.animate(xAxisDuration: 1.0, yAxisDuration: 1.0, easingOption: .linear)
+extension MainVC{
+    private func postRecordData(){
+        swimmingViewModel
+            .swimmingSubject
+            .bind(onNext: { [weak self] workoutList in
+                guard let self = self else { return }
                 
-                if currentState == .week {
-                    cell.titleLabel.text = "WEEKLY RECORD"
-                    cell.titleLabel.addCharacterSpacing(kernValue: 2)
-                } else {
-                    cell.titleLabel.text = "MONTHLY RECORD"
-                    cell.titleLabel.addCharacterSpacing(kernValue: 2)
+                print("workoutList.count------------\(workoutList.count)---------------")
+                workoutList.forEach{
+                    print("\($0.display())")
+                    print("count = \($0.recordLabsList.count)")
                 }
-                return cell
-            case 3:
-                guard let cell = tableView.dequeueReusableCell(withIdentifier: DetailTVC.identifier) as? DetailTVC else { return UITableViewCell() }
+                print("----------------------------")
                 
-                if currentState == .week {
-                    cell.titleLabel.text = "WEEKLY OVERVIEW"
-                    cell.titleLabel.addCharacterSpacing(kernValue: 2)
-                } else {
-                    cell.titleLabel.text = "MONTHLY OVERVIEW"
-                    cell.titleLabel.addCharacterSpacing(kernValue: 2)
+                self.storage.dispatchRecord(workoutList: workoutList) {
+                    print("success")
                 }
-                
-                return cell
-            default:
-                let cell = UITableViewCell(frame: .zero)
-                cell.backgroundColor = .upuhBackground
-                cell.selectionStyle = .none
-                return cell
-            }
-        case .routine:
-            guard let cell = tableView.dequeueReusableCell(withIdentifier: RoutineTVC.identifier) as? RoutineTVC else { return UITableViewCell()}
-            cell.setContentData(overview: routineOverViewList[indexPath.row])
-            return cell
-        }
+            }).disposed(by: disposeBag)
     }
 }
 
@@ -388,157 +223,5 @@ extension MainVC: UITableViewDelegate {
             cardView.cardViewState = .normal
             decideTopConstraint(of: .normal)
         }
-    }
-}
-
-// MARK: - SelectedRange Delegate
-extension MainVC: SelectedRangeDelegate {
-    func didClickedRangeButton() {
-        cardView.cardViewState = .normal
-        decideTopConstraint(of: .normal)
-        
-        guard let vc = storyboard?.instantiateViewController(withIdentifier: "SelectedRangeVC") as? SelectedRangeVC else { return }
-        vc.modalPresentationStyle = .overCurrentContext
-        vc.modalTransitionStyle = .crossDissolve
-        
-        vc.dayData = { year, month, day in
-            print("dayData : \(year) \(month) \(day)")
-            let transYear = year[year.index(year.startIndex, offsetBy: 2)..<year.endIndex]
-            let transMonth = (month.count == 1) ? "0\(month)" : month
-            let transDay = (day.count == 1) ? "0\(day)" : day
-            self.dateText = "\(transYear)/\(transMonth)/\(transDay)"
-            self.cardView.dateText = self.dateText
-            self.currentState = (self.dateText == self.dateformatter.string(from: Date())) ? .base : .day
-            self.strokeState = .none
-            
-            // fetch day response
-            self.rangeTexts[0] = "\(year)-\(transMonth)-\(transDay)"
-            self.storage.dispatchDayRecord(date: self.rangeTexts[0], stroke: "") {
-                print("day Record")
-                
-                self.baseTableView.reloadSections(IndexSet(1...1), with: .fade)
-                self.cardView.currentState = self.currentState
-            }
-        }
-        vc.weekData = { week in
-            print("weekData : \(week)")
-            self.dateText = week
-            self.cardView.dateText = self.dateText
-            self.currentState = .week
-            self.strokeState = .none
-            
-            // fetch week response
-            self.rangeTexts[0] = "2021-10-18"
-            self.rangeTexts[1] = "2021-10-24"
-            
-            self.storage.dispatchWeekRecord(startDate: self.rangeTexts[0],
-                                            endDate: self.rangeTexts[1],
-                                            stroke: "") {
-                print("Week Record")
-                
-                self.baseTableView.reloadSections(IndexSet(1...1), with: .automatic)
-                self.cardView.currentState = self.currentState
-            }
-        }
-        vc.monthData = { year, month in
-            print("monthData : \(year) \(month)")
-            let transMonth = (month.count == 1) ? "0\(month)" : month
-            self.dateText = "\(year)/\(transMonth)"
-            self.cardView.dateText = self.dateText
-            self.currentState = .month
-            self.strokeState = .none
-            
-            // fetch month response
-            self.rangeTexts[0] = "\(year)-\(transMonth)"
-            self.storage.dispatchDayRecord(date: self.rangeTexts[0], stroke: "") {
-                print("month Record")
-                
-                self.baseTableView.reloadSections(IndexSet(1...1), with: .automatic)
-                self.cardView.currentState = self.currentState
-            }
-        }
-        
-        present(vc, animated: true, completion: nil)
-    }
-    
-    func didClickedStrokeButton(indexPath: Int = 0) {
-        cardView.cardViewState = .normal
-        decideTopConstraint(of: .normal)
-        
-        guard let vc = storyboard?.instantiateViewController(withIdentifier: "SelectedStrokeVC") as? SelectedStrokeVC else { return }
-        vc.modalPresentationStyle = .overCurrentContext
-        vc.modalTransitionStyle = .crossDissolve
-        vc.style = self.strokeState
-        vc.strokeData = { style in
-            print(style)
-            self.strokeState = style
-            
-            switch self.currentState {
-            case .week:
-                self.storage.dispatchWeekRecord(startDate: self.rangeTexts[0],
-                                                endDate: self.rangeTexts[1],
-                                                stroke: style.rawValue) {
-                    print("Week Record with stroke")
-                    
-                    self.baseTableView.reloadSections(IndexSet(1...1), with: .automatic)
-                }
-            case .month:
-                self.storage.dispatchMonthRecord(date: self.rangeTexts[0],
-                                                 stroke: style.rawValue) {
-                    print("Month Record with stroke")
-                    
-                    self.baseTableView.reloadSections(IndexSet(1...1), with: .automatic)
-                }
-            default:
-                break
-            }
-        }
-        
-        present(vc, animated: true, completion: nil)
-    }
-}
-
-// MARK: - HealthKit
-extension MainVC {
-    private func authorizeHealthKit() {
-        HealthKitSetupAssistant.authorizeHealthKitAtSwimming { (authorized, error) in
-            guard authorized else {
-                let baseMessage = "HealthKit Authorization Failed"
-                if let error = error {
-                    print("\(baseMessage). Reason: \(error.localizedDescription)")
-                } else {
-                    print(baseMessage)
-                }
-                return
-            }
-            print("HealthKit Successfully Authorized.")
-            self.postRecordData()
-            self.swimmingViewModel.initSwimmingData()
-            DispatchQueue.main.asyncAfter(deadline: .now() + 3){
-                self.swimmingViewModel.refineSwimmingDataForServer()
-                print("refineSwimmingDataForServer.")
-            }
-        }
-    }
-}
-
-extension MainVC{
-    private func postRecordData(){
-        swimmingViewModel
-            .swimmingSubject
-            .bind(onNext: { [weak self] workoutList in
-                guard let self = self else { return }
-                
-                print("workoutList.count------------\(workoutList.count)---------------")
-                workoutList.forEach{
-                    print("\($0.display())")
-                    print("count = \($0.recordLabsList.count)")
-                }
-                print("----------------------------")
-                
-                self.storage.dispatchRecord(workoutList: workoutList) {
-                    print("success")
-                }
-            }).disposed(by: disposeBag)
     }
 }
